refactor: share one reasonOf between the CLI and the pipeline
The CLI and the conversion pipeline each had a private copy of the
helper that flattens a caught error to its message, and the CLI's copy
kept the message's line breaks. Both now use the one in src/error.ts,
which collapses the whitespace so the message fits on one line of
another message.
